Fix election page quirks. Fixes #77, fixes #86
This simplifies the templating and Javascript around the elections.

We ensure the candidate order is stable for each individual voter (but
varies randomly from voter to voter to avoid bias), which allows us
to update both the ballot and the candidate list on every state change.
This eliminates the synchronization bugs the old code suffered from by
always quickly converging to a sane state.

The polling frequency is also reduced to reduce load on the system; it
only matters if the user has multiple browsers open.

This also fixes the final issue keeping the mobile-device-support issue
open.